Move motion_service_sim into its own package
g1_bringup held the full implementation, its unit tests and the ONNX weights. A
bringup package should hold launch files, config and scripts.

Everything moved exists only because unitree_mujoco stops at the low-level
device, so it is one deletion boundary. The launch suites stay behind: they all
include sim.launch.py, so hosting them there would be a dependency cycle.
