+++ workspace/src/g1_motion_service_sim/package.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>g1_motion_service_sim</name>
+  <version>0.1.0</version>
+  <description>
+    SIM-ONLY stand-in for the G1's onboard motion service. unitree_mujoco emulates
+    only the low-level device, so nothing serves the weight-blended /arm_sdk interface,
+    nothing answers /api/sport/* and nothing commands the legs. This node fills all
+    three gaps, sim-side only: it blends /arm_sdk onto /lowcmd, answers the LocoClient
+    wire protocol, and runs the RL walking policy that balances the robot. Launched
+    exclusively by g1_bringup's sim.launch.py. See README.
+  </description>
+  <maintainer email="gupta.adyansh@gmail.com">Adyansh</maintainer>
+  <license>BSD-3-Clause</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
+  <depend>rclcpp</depend>
+  <depend>unitree_hg</depend>
+  <depend>unitree_api</depend>
+  <depend>unitree_go</depend>
+  <depend>g1_hardware_interface</depend>
+  <depend>ament_index_cpp</depend>
+  <depend>sensor_msgs</depend>
+
+  <!-- nlohmann_json and onnxruntime are deliberately not declared: both come from
+       .devcontainer/Dockerfile rather than rosdep, matching g1_locomotion. -->
+
+  <!-- Lint checks run via workspace-root CTests (see CMakeLists.txt). -->
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_cmake_lint_cmake</test_depend>
+  <test_depend>ament_cmake_xmllint</test_depend>
+  <test_depend>ament_cmake_gmock</test_depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
