+++ workspace/src/g1_motion_service_sim/test/test_blend_math.cpp
@@ -4,9 +4,9 @@
  */
 #include <gmock/gmock.h>
 
-#include "g1_bringup/blend_math.hpp"
+#include "g1_motion_service_sim/blend_math.hpp"
 
-namespace g1_bringup
+namespace g1_motion_service_sim
 {
 namespace
 {
@@ -126,4 +126,4 @@ TEST(StepEffectiveWeight, ResultIsAlwaysClampedToUnitRange)
 }
 
 }  // namespace
-}  // namespace g1_bringup
+}  // namespace g1_motion_service_sim
